telegram/pyproject.toml
[tool.poetry]
name = "telegram"
version = "0.1.0"
description = ""
authors = ["yilbegan <w2v@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
aiogram = "^2.9"
aiohttp = {extras = ["speedups"], version = "^3.6.2"}
environs = "^8.0.0"
Pillow = "^7.1.2"
aiofile = "^1.5.2"
tortoise-orm = "^0.16.13"
asyncpg = "^0.20.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
aiohttp_socks = "^0.3.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
